feat: improve verbose output readability
Redesign the --verbose impacted objects output to help users more easily
scan, correlate, and act on diagnostic results.

Application-wide changes:

- Extract table rendering logic from command_options.go into
  output_table.go, separating rendering concerns from CLI options,
  validation, and sorting
- Render impacted objects in a bordered 5-column table (STATUS, KIND,
  GROUP, CHECK, IMPACT) matching the summary table's visual style
- Sort verbose output entries using the same order as the summary table
  (group -> kind -> impact -> check) for easy cross-reference
- Size table width dynamically to accommodate the longest detail line,
  preventing truncation of image paths and other verbose content
- Handle ANSI color codes and Unicode symbols correctly when computing
  column alignment
- Add nil-Result guards in collectSortedRows and buildVerboseRows to
  prevent nil-pointer dereferences
- Add ResourceType.CRDFQN() method and AnnotationResourceCRDName
  annotation so impacted object CRD names (e.g. notebooks.kubeflow.org)
  are resolved from authoritative ResourceType definitions rather than
  string derivation

Notebook-specific changes:

- Add NotebookVerboseFormatter, a shared formatter embedded in notebook
  checks, providing consistent namespace-grouped output with requester
  info and CRD-qualified resource names
- Restructure impacted-workloads verbose output to group notebooks by
  namespace within each image section, replacing flat namespace/name
  listings with hierarchical output that makes it easier to identify
  which team owns affected workloads
- Sort image groups so incompatible images appear before custom images,
  with secondary sort by image reference for deterministic output
- Sort notebook names within each namespace for stable output across
  runs
- Guard against nil Annotations map in setImpactedObjects, matching the
  defensive pattern used by SetImpactedObjects and AddImpactedObjects
