+++ pkg/lint/checks/workloads/notebook/accelerator_migration.go
@@ -18,6 +18,7 @@ import (
 // that will be auto-migrated to HardwareProfiles (infrastructure.opendatahub.io) during RHOAI 3.x upgrade.
 type AcceleratorMigrationCheck struct {
 	check.BaseCheck
+	NotebookVerboseFormatter
 }
 
 func NewAcceleratorMigrationCheck() *AcceleratorMigrationCheck {
