+++ pkg/lint/checks/workloads/notebook/container_name.go
@@ -16,6 +16,7 @@ import (
 // Dashboard-managed annotations (accelerator profile or size selection) are checked.
 type ContainerNameCheck struct {
 	check.BaseCheck
+	NotebookVerboseFormatter
 }
 
 // NewContainerNameCheck creates a new ContainerNameCheck.
